% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rradex.R
\docType{package}
\name{rradex-package}
\alias{rradex-package}
\title{A package for performing radiative transfer calculations.}
\description{
This is an R-interface to
[RADEX](http://home.strw.leidenuniv.nl/~moldata/radex.html), a
statistical equilibrium radiative transfer code, made available for
public use as part of the Leiden Atomic and Molecular Database
(LAMDA). RADEX is a one-dimensional non-LTE radiative transfer code,
that uses the escape probability formulation assuming an isothermal and
homogeneous medium without large-scale velocity fields. RADEX is
comparable to the LVG method and provides a useful tool in rapidly
analyzing a large set of observational data providing constraints on
physical conditions, such as density and kinetic temperature.
}
\details{
The formalism adopted in RADEX is summarized in Van der Tak, F.F.S. et
al.  2007, A&A 468, 627.

Here is an example of what a short session may look like. It's basically
the example given near the end of the RADEX manual (download at
http://www.strw.leidenuniv.nl/~moldata/radex_manual.pdf).
Note, that the various columns of the results table will be immediately
available as vectors in R for further calculations, plotting, etc:
}
\examples{
library(rradex)
\dontrun{
GHz <- 1.0e9
MHz <- 1.0e6
au <- 1.4959787e+11      # astronomical unit in m
pc <- 3.0856776e+16      # parsec in m
Tkin <- 10.0             # kinetic temperature
Tbg <- 2.73              # background temperature
d <- 3000*au             # diameter of source
f.CO <- 8.0e-5           # fractional abundance of CO
nH2 <- 3.0e3             # H2 density
Ncol <- nH2*d*100*f.CO   # CO column density

f <- 115.0*GHz           # frequency of line to be observed
dv <- 1.0                # required velocity resolution
df <- f*dv/299792.46/MHz # ... same in frequency

l = list(partner="H2", density=nH2)                        # collisions with H2 ...
# l = list(partner=c("o-H2","p-H2"), density=c(3,1)*nH2/4) # ... or with explicit o/p ratio
# give filename with full path to molecular data file
molfile <- "/path/to/co.dat"

# now call the RADEX routine
result <- radex(molfile, freqs=c(100,500), Tkin=Tkin, Ncol=Ncol, collisions=l)
# print and plot results
print(result)
plot(result$GHz, result$T.R, xlim=c(100,500), type='h', col='blue', lwd=4,
     xlab="frequency [GHz]", ylab=expression(T[R]))
}
}
\author{
Michael Olberg, \email{michael.olberg@chalmers.se}
}

