% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loaddata}
\alias{loaddata}
\title{Load a molecular data file from the Leiden Atomic and Molecular
Database (LAMDA) database for easy inspection of levels and
transitions.}
\usage{
loaddata(molfile)
}
\arguments{
\item{molfile}{filename of molecular data file, including full path}
}
\value{
a list with components "molecule" (molecule name), "amass"
    (molecular mass in a.m.u.), "levels" (energy levels),
    "transitions" (radiative transitions) and "collisions"
    (collision rates)
}
\description{
Load a molecular data file from the Leiden Atomic and Molecular
Database (LAMDA) database for easy inspection of levels and
transitions.
}
\examples{
library(rradex)
\dontrun{
result <- loaddata("data/hco+.dat")
print(result)
}
}
\keyword{radiative}
\keyword{transfer}
